rstructure and minor fixes

+++ bai3/bai15.h
@@ -2,13 +2,13 @@
 
 #pragma once
 double sn15(int n) {
-  double result = 0;
-  int sum = 0;
+    double result = 0;
+    int sum = 0;
 
-  for (float i = 1.0; i <= n; ++i) {
-    sum += i;
-    result += (float)1 / (sum);
-  }
+    for (int i = 1.0; i <= n; ++i) {
+        sum += i;
+        result += 1.0 / (sum);
+    }
 
-  return result;
+    return result;
 }
